Move connectToGitHub to its own file
